+++ src/modules/ethEventsScraper/service_layer/ethEventsScraper.service.ts
@@ -16,6 +16,7 @@ import { MintingCollection } from '../../nft/domain/minting-collection.entity';
 
 @Injectable()
 export class EthEventsScraperService {
+  private logger = new Logger(EthEventsScraperService.name);
   processing = false;
 
   constructor(
@@ -40,11 +41,18 @@ export class EthEventsScraperService {
 
   @Cron(CronExpression.EVERY_30_SECONDS)
   public async syncCollectionAndNftEvents() {
-    if (this.processing) return;
-    this.processing = true;
-    await this.syncDeployCollectionEvents();
-    await this.syncMintNftEvents();
-    this.processing = false;
+    this.logger.log('start');
+    try {
+      if (this.processing) return;
+      this.processing = true;
+      await this.syncDeployCollectionEvents();
+      await this.syncMintNftEvents();
+      this.processing = false;
+    } catch (e) {
+      this.processing = false;
+      console.log(e);
+    }
+    this.logger.log('end');
   }
 
   private async syncDeployCollectionEvents() {
@@ -56,6 +64,7 @@ export class EthEventsScraperService {
       const collection = this.nftCollectionRepository.create();
       collection.txHash = event.tx_hash;
       collection.address = event.contract_address?.toLowerCase();
+      collection.owner = event.owner.toLowerCase();
       collection.name = event.token_name;
       collection.symbol = event.token_symbol;
 
@@ -74,24 +83,16 @@ export class EthEventsScraperService {
 
   private async syncMintNftEvents() {
     const events = await this.createNftEventRepository.find({ where: { processed: false } });
-    const txHashAndEventsMap = events.reduce((acc, event) => {
-      const prevEventsForTxHash = acc[event.tx_hash] || [];
-      return {
-        ...acc,
-        [event.tx_hash]: [...prevEventsForTxHash, event],
-      };
-    }, {} as Record<string, MintedNftEvent[]>);
+    const tokenUriEventsMap = this.mapTokenUriToEvents(events);
 
-    const txHashes = Object.keys(txHashAndEventsMap);
-
-    for (const txHash of txHashes) {
+    for (const tokenUri of Object.keys(tokenUriEventsMap)) {
       const editionUUID = customAlphabet('1234567890abcdefghijklmnopqrstuvwxyz', 10)();
-      const events = txHashAndEventsMap[txHash];
+      const events = tokenUriEventsMap[tokenUri];
 
       for (const event of events) {
         const response = await this.httpService.get(event.token_uri).toPromise();
         const artworkType = (response.data.image_url as string).split(/[.]+/);
-        const user = await this.userRepository.findOne({ where: { address: event.receiver } });
+        const user = await this.userRepository.findOne({ where: { address: event.receiver.toLowerCase() } });
         const collection = await this.nftCollectionRepository.findOne({
           where: { address: event.contract_address.toLowerCase() },
         });
@@ -100,7 +101,7 @@ export class EthEventsScraperService {
           const nft = this.nftRepository.create();
           nft.userId = user.id;
           nft.collectionId = collection.id;
-          nft.txHash = txHash;
+          nft.txHash = event.tx_hash;
           nft.editionUUID = editionUUID;
           nft.name = response.data.name as string;
           nft.description = response.data.description as string;
@@ -120,11 +121,22 @@ export class EthEventsScraperService {
           await this.nftRepository.save(nft);
         }
       }
-      await this.savedNftRepository.softDelete({ txHash });
+      await this.savedNftRepository.softDelete({ tokenUri });
     }
   }
 
-  //Todo: search for auctions with txHash and onChain flag false
+  private mapTokenUriToEvents(events: MintedNftEvent[]) {
+    const tokenUriEventsMap = events.reduce((acc, event) => {
+      const prevEventsForTokenUri = acc[event.token_uri] || [];
+      return {
+        ...acc,
+        [event.token_uri]: [...prevEventsForTokenUri, event],
+      };
+    }, {} as Record<string, MintedNftEvent[]>);
+    return tokenUriEventsMap;
+  }
+
+//Todo: search for auctions with txHash and onChain flag false
   async syncCreateAuctionEvents() {}
 
   //We might need to rethink how things are stored in the database, don't know if the jsonb encoded data is ok